backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.12
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# OpenAI (for Phase 4)
openai==1.57.0

# Utilities
python-dateutil==2.9.0

# Testing (optional for later)
pytest==8.3.4
httpx==0.28.1